+++ ai-core/source/uim/platform/ai_core/application/use_cases/manage/configurations.d
@@ -90,7 +90,7 @@ class ManageConfigurationsUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long count(ResourceGroupId rgId) {
+  size_t count(ResourceGroupId rgId) {
     return repo.countByResourceGroup(rgId);
   }
 }
