+++ ai-core/source/uim/platform/ai_core/application/use_cases/manage/deployments.d
@@ -120,7 +120,7 @@ class ManageDeploymentsUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long count(ResourceGroupId rgId) {
+  size_t count(ResourceGroupId rgId) {
     return deplRepo.countByResourceGroup(rgId);
   }
 }
