+++ ai-core/source/uim/platform/ai_core/application/use_cases/manage/executions.d
@@ -108,7 +108,7 @@ class ManageExecutionsUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long count(ResourceGroupId rgId) {
+  size_t count(ResourceGroupId rgId) {
     return execRepo.countByResourceGroup(rgId);
   }
 }
