+++ ai-core/source/uim/platform/ai_core/application/use_cases/manage/scenarios.d
@@ -66,7 +66,7 @@ class ManageScenariosUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long count(ResourceGroupId rgId) {
+  size_t count(ResourceGroupId rgId) {
     return repo.countByResourceGroup(rgId);
   }
 }
