+++ ai-core/source/uim/platform/ai_core/infrastructure/persistence/memory/configurations.d
@@ -53,7 +53,7 @@ class MemoryConfigurationRepository : ConfigurationRepository {
     }
   }
 
-  long countByResourceGroup(ResourceGroupId rgId) {
+  size_t countByResourceGroup(ResourceGroupId rgId) {
     if (auto rg = rgId in store)
       return cast(long)(*rg).length;
     return 0;
