+++ datasphere/source/uim/platform/datasphere/infrastructure/persistence/memory/remote_table_repo.d
@@ -58,7 +58,7 @@ class MemoryRemoteTableRepository : RemoteTableRepository {
     }
   }
 
-  long countBySpace(SpaceId spaceId) {
+  size_t countBySpace(SpaceId spaceId) {
     if (auto sp = spaceId in store)
       return cast(long)(*sp).length;
     return 0;
