+++ datasphere/source/uim/platform/datasphere/infrastructure/persistence/memory/view_repo.d
@@ -64,7 +64,7 @@ class MemoryViewRepository : ViewRepository {
     }
   }
 
-  long countBySpace(SpaceId spaceId) {
+  size_t countBySpace(SpaceId spaceId) {
     if (auto sp = spaceId in store)
       return cast(long)(*sp).length;
     return 0;
