+++ hana/source/uim/platform/hana/application/usecases/manage/schemas.d
@@ -79,7 +79,7 @@ class ManageSchemasUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long count(TenantId tenantId) {
+  size_t count(TenantId tenantId) {
     return repo.countByTenant(tenantId);
   }
 }
